+++ models/BarTicket.java
