+++ models/TicketStatus.java
@@ -1,0 +1,15 @@
+package models;
+
+/**
+ * This enumeration describes the base ticket states.
+ * 
+ * Please note this is purposely stubbed until the SRD is finished.
+ * @author Ian Wilhelmsen Last Updated: 2/19/20
+ *
+ */
+public enum TicketStatus {
+    TICKETSTATUS1,
+    TICKETSTATUS2,
+    TICKETSTATUS3
+
+}
